+++ Xqueue/Maileon/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.9.4
+- Fix order.categories and order.product_ids max. 1000 char length at transaction types
+- Add plugin status reporter cron job
+
 # 1.9.3
 - Add getCustomOrderTransactionAttributes() helper method to order confirm transaction